shared_libs/requirements.txt

boto3
botocore
fastapi
fastembed
groq
jsonschema
langchain_community
langchain_community
langchain_text_splitters
langdetect
lxml
mangum
numpy
openai
pandas
protobuf
pydantic
pytest
python-dotenv
python_docx
PyYAML
qdrant_client
Requests
scikit_learn
tqdm
